Correctly initialise router_addr.remote_addr, use correct width for rpacket_header.src_ip and correctly set router copy of socket flags.

+++ src/router.h
@@ -105,7 +105,7 @@ struct rclient {
 };
 
 struct rpacket_header {
-	uint8_t src_ipaddr;
+	uint32_t src_ipaddr;
 	char spare[20];
 } __attribute__((__packed__));
